scripts/video-editing/requirements.txt
faster-whisper>=1.1.1
jieba>=0.42.1
Pillow>=10.4.0
pytest>=7.4.4
